feat(core, angular): Show errors in chat and completion (#154)
This PR wires in support for:
* showing errors in chat with a retry option
* detecting when service is down or borked (via using up retries) so the
async background things (i.e. input completion) can alert a user when
that feature is not working

Both the above are working in the Angular package and app for chat and
completion.

Subsequent PRs will add service outage support for:
* prediction
* structured completion
* similar support in the React package and app.

It felt like it made sense to:
* add a 'working' indicator to the chat, since when you click 'retry' on
an error, nothing obvious was happening.