Bug(170) forego any of wrappers when not streaming (#180)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [ ] The commit message follows our guidelines:
https://github.com/liveloveapp/hashbrown/blob/main/CONTRIBUTING.md#commit
- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

```
[X] Bugfix
[ ] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Build related changes
[ ] CI related changes
[ ] Documentation content changes
[ ] Other... Please describe:
```

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->
LLMs seem to struggle with the discriminator wrapper objects we
introduced for types in an anyOf.

Closes #170

## What is the new behavior?
When parsing objects, arrays, anyOfs or streaming strings, we still need
the discriminator to know how to handle incomplete objects, so those
remain.

For other types (non-streaming strings, types that cannot stream in the
first place), discriminator objects are no longer added (or when
parsing, removed) from JSONSchema or data.

## Does this PR introduce a breaking change?

```
[ ] Yes
[X] No
```

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information