bugfix(136): Support unwrapping of anyOf type as a direct child of an anyOf (#196)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [ ] The commit message follows our guidelines:
https://github.com/liveloveapp/hashbrown/blob/main/CONTRIBUTING.md#commit
- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

```
[X] Bugfix
[ ] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Build related changes
[ ] CI related changes
[ ] Documentation content changes
[ ] Other... Please describe:
```

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->
A nested anyOf was being parsed as an object, so the wrapper was staying
in the output.

Closes #136

## What is the new behavior?
In parseAnyOf, if the matching schema of a discriminator is itself an
anyOf, the options array is added to the container stack instead of the
AnyOf schema.

That way, the subsequent call to parseAny will route to parseAnyOf
instead of parseObj.

## Does this PR introduce a breaking change?

```
[ ] Yes
[X] No
```

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information