docs(core): remove symbols duplication (#384)
When generating docs for an overloaded method the symbols would be
duplicated. This moves the section above the loop so that it's only
display 1 time.

Closes #311

## PR Checklist

Please check if your PR fulfills the following requirements:

- [x] The commit message follows our guidelines:
https://github.com/liveloveapp/hashbrown/blob/main/CONTRIBUTING.md#commit
- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

```
[ ] Bugfix
[ ] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Build related changes
[ ] CI related changes
[x] Documentation content changes
[ ] Other... Please describe:
```

## What is the current behavior?

The header and summary of the method is rendered for each symbol
declaration.
This causes it to be duplicated for overloaded methods.

Closes #311 

## What is the new behavior?

The symbol summary and header are now only displayed one time on the
docs page.
They have been moved outside of the for loop.

## Does this PR introduce a breaking change?

```
[ ] Yes
[x] No
```

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information