Specify model format in azure library (#286)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [x] The commit message follows our guidelines:
https://github.com/liveloveapp/hashbrown/blob/main/CONTRIBUTING.md#commit
- [x] Tests for the changes have been added (for bug fixes / features)
- [x] Documentation has been added / updated (for bug fixes / features)

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

```
[ ] Bugfix
[ ] Feature
[x] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Build related changes
[ ] CI related changes
[ ] Documentation content changes
[ ] Other... Please describe:
```

## What is the current behavior?

Currently the azure package accepts a generic request type

## What is the new behavior?
Makes the azure request a bit more specific to include the model and
version number instead of just a string. This should help with
intellisense etc.. and hopefully make the format the model must be sent
as more clear


## Does this PR introduce a breaking change?

```
[ ] Yes
[x] No
```