Change deployment configuration to target Cloudflare Pages runtime (#450)
The 'www' project deploys to Cloudflare and uses SSR.

Since it uses SSR, it needs to target Cloudflare's custom Pages runtime
instead of Node. This PR does that.

Also, the Pages runtime does not support dynamic WASM libraries, so this
PR removes Shiki in favor of a plain-text.

For whatever reason, the deployment commands that had previously seemed
to need to run in working directory of the pages projects:
* don't anymore
* weren't deploying worker.js files needed for CloudFlare SSR to work.

So, the PR also simplifies those commands to ones that do correctly
deploy those files.


## PR Checklist

Please check if your PR fulfills the following requirements:

- [ ] The commit message follows our guidelines:
https://github.com/liveloveapp/hashbrown/blob/main/CONTRIBUTING.md#commit
- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

```
[ ] Bugfix
[ ] Feature
[ ] Code style update (formatting, local variables)
[ ] Refactoring (no functional changes, no api changes)
[ ] Build related changes
[ ] CI related changes
[ ] Documentation content changes
[ ] Other... Please describe:
```

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Closes #

## What is the new behavior?

## Does this PR introduce a breaking change?

```
[ ] Yes
[ ] No
```

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information